Cross ISCO 88 to ISCO 08 - n digit

Cross ISCO 88 to ISCO 08 - n digit

isco88_to_isco08_ndigit(
  base,
  isco,
  digits,
  summary = FALSE,
  code_titles = FALSE
)

Arguments

base

Dataframe from any survey including a variable with ISCO 88 codes

isco

character vector containing ISCO 88 codes

digits

Digits number (ranging from 1 to 4) for ISCO codes in input data

summary

If TRUE provides other dataframe counting how many cases where asigned for each ISCO88 code to each ISCO08 code.

Value

The function returns the provided dataframe, adding a new variable with the crosswalk to ISCO 08 codes

Examples

toy_base_peru_isco08 <- isco88_to_isco08_ndigit(toy_base_peru,isco = p505,digits = 3)
#> Warning: The following codes from the provided database were not found in 'crosstable_isco08_isco88' and it was not possible to crosswalk them: c("252", "216", "375", "942", "572", "247", "352", "919", "574", "423", "454", "218", "971", "783", "922", "987", "985", "945", "768", "941", "861", "982", "541", "885", "881", "319", "952", "354", "983", "364", "351", "623", "316", "254", "239", "372", "785", "251", "219", "777", "781", "771", "381", "335", "868", "981", "772", "571", "843", "782", "23", "917", "943", "766", "573", "632", "391", "12", "396", "145", "751", "877", "746", "229", "258", "851", "227", "395", "415", "136", "747", "778", 
#> "842", "361", "356", "765", "961", "564", "792", "864", "875", "953", "451", "562", "927", "852", "581", "217", "379", "21", "374", "269", "367", "776", "268", "377", "737", "238", "362", "462", "794", "872", "261", "761", "442", "773", "944", "453", "876", "984", "951", "436", "226", "267", "836", "883", "795", "531", "973", "461", "366", "271", "455", "236", "139", "791", "144", "844", "918", "924", "13", "137", "127", "873", "452", "745", "394", "355", "616", "262", "443", "767", "923", "349", 
#> "272", "224", "749", "256", "797", "641", "764", "972", "774", "418", "317", "266", "11", "318", "624", "841", "775", "784", "147", "376", "228", "839", "886", "763", "865", "867", "392", "22", "871", "135", "373", "138", "926", "779", "793", "563", "636", "284", "748", "148", "143", "631", "142", "263", "365", "925", "456", "237", "264", "553", "862", "769", "633", "617")