Reclassify ocupational codes to International Standard Classification of Occupations 08

Reclassify ocupational codes to International Standard Classification of Occupations 08

reclassify_to_isco08(
  base,
  variable,
  classif_origin,
  add_skill = F,
  add_major_groups = F,
  code_titles = F,
  summary = F
)

Arguments

base

Dataframe including a variable with occupational codes contained in "available_classifications"

variable

variable containing occupational codes

classif_origin

character vector specifying the Classification system used as input. The supported clasification systems are: "Census2010" (United States), "CNO2017" (Argentina), "SINCO2011" (Mexico), "ISCO88", "ISCO88_3digits" and "ISCO08". If variable is already classified with ISCO08, setting "ISCO08" to this parameter allows you use the function to add new variables with major groups and skill levels.

add_skill

If TRUE adds a new variable with the occupation skill level based on ISCO 08 skill levels. The new skill_level variable is a factor variable containg the levels 'Low', 'Medium' and 'High'.

add_major_groups

If TRUE adds a new variable with the major groups of each occupation based on the 1-digit ISCo-08 classification structure. The major gruops structure has 9 levels, where level 1 means the highest skill level needed to perform the job and level 9 refers ti the lowest skill level.

code_titles

If TRUE adds classification titles besides from codes.

summary

If TRUE provides other dataframe counting how many cases where asigned for each Census 08 code to each ISCO88 code.

Value

The function returns the provided dataframe, adding a new variable with the crosswalk to ISCO 08 codes

Details

disclaimer: This script uses crosswalks provided by different statistical offices arround the world. It is not an official product of any of them

Examples

USA_database_with_isco08 <- reclassify_to_isco08(toy_base_ipums_cps_2018, OCC2010, classif_origin="Census2010")
#> Warning: The following codes from the provided database were not found in 'crosstable_census2010_soc2010_isco08' and it was not possible to crosswalk them:  c("1100", "730", "620", "3650", "2150", "3530", "3130", "2140", "3410", "9100", "130", "1960", "560")
MEX_database_with_isco08 <- reclassify_to_isco08(toy_base_mexico, p3, classif_origin="SINCO2011")
#> Warning: The following codes from the provided database were not found in 'crosstable_sinco2011_isco08' and it was not possible to crosswalk them:  NA
ARG_database_with_isco08 <- reclassify_to_isco08(toy_base_eph_argentina, PP04D_COD, classif_origin="CNO2001")
#> Warning: The following codes from the provided database were not found in 'crosstable_cno2001_isco08' and it was not possible to crosswalk them: c(NA, "92333", "43331", "45331", "71331", "99998", "30132", "81339", "30339", "30319", "99997", "42339", "99991", "30123", "99999", "54333", "31333", "11331", "46333", "61331", "20131", "11131", "33123", "82131", "57122", "43332", "44331", "90333", "41331", "54322", "43131")