Reclassify ocupational codes to International Standard Classification of Occupations 08

Reclassify ocupational codes to International Standard Classification of Occupations 08

reclassify_to_isco08(
  base,
  variable,
  classif_origin,
  add_complexity = F,
  code_titles = F,
  summary = F
)

Arguments

base
Dataframe including a variable with occupational codes contained in "available_classifications"

variable
variable containing occupational codes

classif_origin
character vector specifying the Classification. "Census2010" (United States), "CNO2017" (Argentina), "SINCO2011" (Mexico) are supported.

add_complexity
If TRUE adds a new variable with the occupation complexity level based on ISCO 08 skill levels. The new complexity_level variable is a factor variable containg the levels 'Low', 'Medium' and 'High'.

code_titles
If TRUE adds classification titles besides from codes.

summary
If TRUE provides other dataframe counting how many cases where asigned for each Census 08 code to each ISCO88 code.

Value

The function returns the provided dataframe, adding a new variable with the crosswalk to ISCO 08 codes

Details

disclaimer: This script uses crosswalks provided by different statistical offices arround the world. It is not an official product of any of them

Examples